improved ui and structure of order master data
